+++ analysis-master/tra_analysis/regression_old.py
